requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
google-genai>=1.14.0
google-cloud-vision>=3.7.0
google-cloud-speech>=2.27.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
slack-sdk>=3.0.0
aiohttp>=3.9.0
httpx>=0.28.0
pytest>=8.3.0
typing_extensions>=4.12.0
